before migrating rc7 to latest rc

+++ plzoo-workspace/imp.lang.truffle/interpreter-comp/interpreter/imp/interpreter/imp/IntValue.java
@@ -4,7 +4,7 @@ import com.oracle.truffle.api.nodes.NodeInterface;
 import org.eclipse.emf.ecore.EObject;
 
 public interface IntValue extends EObject, NodeInterface, Value {
-  long getValue();
+  int getValue();
 
-  void setValue(long value);
+  void setValue(int value);
 }
